+++ chess-ai/ttable.py
